+++ website/src/components/ui/form-label.tsx
@@ -2,5 +2,5 @@ import { ark } from '@ark-ui/react/factory'
 import { styled, type HTMLStyledProps } from 'styled-system/jsx'
 import { label } from 'styled-system/recipes'
 
-export const Label = styled(ark.label, label)
-export interface LabelProps extends HTMLStyledProps<typeof Label> {}
+export const FormLabel = styled(ark.label, label)
+export interface FormLabelProps extends HTMLStyledProps<typeof FormLabel> {}